Extract shared alerting components from the cc-alerting pages
